pi-service/requirements.txt
websockets>=12.0
aiohttp>=3.9
gpiozero>=2.0
python-dotenv>=1.0
RPi.GPIO
